apply review comments 3
