+++ components/external-services-mock/internal/subscription/handler.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"strings"
 
@@ -15,23 +14,16 @@ import (
 	"github.com/tidwall/sjson"
 )
 
-const compassURL = "https://github.com/kyma-incubator/compass"
-
 type handler struct {
 	httpClient     *http.Client
 	tenantConfig   Config
 	providerConfig ProviderConfig
-	xsappnameClone string
 	jobID          string
 }
 
 type JobStatus struct {
 	ID    string `json:"id"`
 	State string `json:"state"`
-}
-
-type Dependency struct {
-	Xsappname string `json:"xsappname"`
 }
 
 func NewHandler(httpClient *http.Client, tenantConfig Config, providerConfig ProviderConfig, jobID string) *handler {
@@ -106,81 +98,6 @@ func (h *handler) JobStatus(writer http.ResponseWriter, r *http.Request) {
 		return
 	}
 	log.C(ctx).Info("Successfully handled subscription job status request")
-}
-
-func (h *handler) OnSubscription(writer http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	log.C(ctx).Info("Handling on subscription request...")
-
-	if r.Method != http.MethodPut && r.Method != http.MethodDelete {
-		log.C(ctx).Errorf("expected %s or %s method but got: %s", http.MethodPut, http.MethodDelete, r.Method)
-		writer.WriteHeader(http.StatusMethodNotAllowed)
-		return
-	}
-
-	writer.Header().Set("Content-Type", "text/plain")
-	writer.WriteHeader(http.StatusOK)
-	if _, err := writer.Write([]byte(compassURL)); err != nil {
-		log.C(ctx).Errorf("while writing response: %s", err.Error())
-		httphelpers.WriteError(writer, errors.Wrap(err, "while writing response"), http.StatusInternalServerError)
-		return
-	}
-	log.C(ctx).Info("Successfully handled on subscription request")
-}
-
-func (h *handler) DependenciesConfigure(writer http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	log.C(ctx).Info("Configuring subscription dependency...")
-	if r.Method != http.MethodPost {
-		log.C(ctx).Errorf("expected %s method but got: %s", http.MethodPost, r.Method)
-		writer.WriteHeader(http.StatusMethodNotAllowed)
-		return
-	}
-
-	body, err := ioutil.ReadAll(r.Body)
-	if err != nil {
-		log.C(ctx).Errorf("while reading request body: %s", err.Error())
-		httphelpers.WriteError(writer, errors.Wrap(err, "while reading request body"), http.StatusInternalServerError)
-		return
-	}
-
-	if string(body) == "" {
-		log.C(ctx).Error("The request body is empty")
-		httphelpers.WriteError(writer, errors.New("The request body is empty"), http.StatusInternalServerError)
-		return
-	}
-
-	h.xsappnameClone = string(body)
-	writer.Header().Set("Content-Type", "application/json")
-	writer.WriteHeader(http.StatusOK)
-	if _, err := writer.Write(body); err != nil {
-		log.C(ctx).Errorf("while writing response: %s", err.Error())
-		httphelpers.WriteError(writer, errors.Wrap(err, "while writing response"), http.StatusInternalServerError)
-		return
-	}
-	log.C(ctx).Infof("Successfully configured subscription dependency: %s", h.xsappnameClone)
-}
-
-func (h *handler) Dependencies(writer http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-	log.C(ctx).Info("Handling dependency request...")
-
-	deps := []*Dependency{{Xsappname: h.xsappnameClone}}
-	depsMarshalled, err := json.Marshal(deps)
-	if err != nil {
-		log.C(ctx).Errorf("while marshalling subscription dependencies: %s", err.Error())
-		httphelpers.WriteError(writer, errors.Wrap(err, "while marshalling subscription dependencies"), http.StatusInternalServerError)
-		return
-	}
-
-	writer.Header().Set("Content-Type", "application/json")
-	writer.WriteHeader(http.StatusOK)
-	if _, err := writer.Write(depsMarshalled); err != nil {
-		log.C(ctx).Errorf("while writing response: %s", err.Error())
-		httphelpers.WriteError(writer, errors.Wrap(err, "while writing response"), http.StatusInternalServerError)
-		return
-	}
-	log.C(ctx).Info("Successfully handled dependency request")
 }
 
 func (h *handler) executeSubscriptionRequest(r *http.Request, httpMethod string) (int, error) {
